Исправление типов данных.

+++ src/main/java/AccountData.java
@@ -2,15 +2,13 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @AllArgsConstructor
 public class AccountData {
-    private Long account;
+    private long account;
 
     private String name;
 
-    private BigDecimal value;
+    private double value;
 }
